added basic reporting model and methods

+++ db/migrate/20090714132448_create_hosts.rb
@@ -15,7 +15,7 @@ class CreateHosts < ActiveRecord::Migration
     add_column :hosts, :root_pass, :string, :limit => 64
     add_column :hosts, :serial, :string, :limit => 12
     add_column :hosts, :puppetmaster, :string
-    add_column :hosts, :services, :string
+    add_column :hosts, :puppet_status, :integer,  :null => false, :default => 0
 
     add_column :hosts, :domain_id, :integer
     add_column :hosts, :architecture_id, :integer
@@ -34,6 +34,14 @@ class CreateHosts < ActiveRecord::Migration
   end
 
   def self.down
-    drop_table :hosts
+    # we are using storeconfigs
+    if Puppet.settings.instance_variable_get(:@values)[:puppetmasterd][:storeconfigs]
+      remove_columns :hosts, :mac, :sp_mac, :sp_ip, :sp_name, :root_pass, :serial,
+        :puppetmaster, :puppet_status, :domain_id, :architecture_id, :operatingsystem_id,
+        :environment_id, :subnet_id, :sp_subnet_id, :ptable_id, :hosttype_id,
+        :media_id, :build, :comment, :disk, :installed_at
+    else
+      drop_table :hosts
+    end
   end
 end
